Relocate store/prodDef functions
